refactor!: move release-notes helper to its own separate module"
