+++ test/github-release.ts
@@ -454,69 +454,4 @@ describe('GitHubRelease', () => {
       requests.done();
     });
   });
-
-  describe('extractLatestReleaseNotes', () => {
-    it('handles CHANGELOG with old and new format entries', () => {
-      const changelogContent = readFileSync(
-        resolve(fixturesPath, './CHANGELOG-old-new.md'),
-        'utf8'
-      ).replace(/\r\n/g, '\n');
-      const latestReleaseNotes = GitHubRelease.extractLatestReleaseNotes(
-        changelogContent,
-        'v1.0.0'
-      );
-      snapshot(latestReleaseNotes);
-    });
-
-    it('handles CHANGELOG with old format entries', () => {
-      const changelogContent = readFileSync(
-        resolve(fixturesPath, './CHANGELOG-old.md'),
-        'utf8'
-      ).replace(/\r\n/g, '\n');
-      const latestReleaseNotes = GitHubRelease.extractLatestReleaseNotes(
-        changelogContent,
-        'v2.1.0'
-      );
-      snapshot(latestReleaseNotes);
-    });
-
-    it('handles CHANGELOG with new format entries', () => {
-      const changelogContent = readFileSync(
-        resolve(fixturesPath, './CHANGELOG-new.md'),
-        'utf8'
-      ).replace(/\r\n/g, '\n');
-      const latestReleaseNotes = GitHubRelease.extractLatestReleaseNotes(
-        changelogContent,
-        'v1.2.0'
-      );
-      snapshot(latestReleaseNotes);
-    });
-
-    // see: https://github.com/googleapis/release-please/issues/140
-    it('extracts appropriate release notes when prior release is patch', () => {
-      const changelogContent = readFileSync(
-        resolve(fixturesPath, './CHANGELOG-bug-140.md'),
-        'utf8'
-      ).replace(/\r\n/g, '\n');
-      const latestReleaseNotes = GitHubRelease.extractLatestReleaseNotes(
-        changelogContent,
-        'v5.0.0'
-      );
-      snapshot(latestReleaseNotes);
-    });
-
-    describe('php-yoshi', () => {
-      it('extracts appropriate release notes, when multiple packages updated', () => {
-        const changelogContent = readFileSync(
-          resolve(fixturesPath, './CHANGELOG-php-yoshi.md'),
-          'utf8'
-        ).replace(/\r\n/g, '\n');
-        const latestReleaseNotes = GitHubRelease.extractLatestReleaseNotes(
-          changelogContent,
-          'v0.105.0'
-        );
-        snapshot(latestReleaseNotes);
-      });
-    });
-  });
 });
